Implementação

+++ file_manager/main.py
@@ -1,6 +1,6 @@
 import os
 import platform
-import sys
+#import sys
 from pathlib import Path
 import time
 
@@ -26,14 +26,21 @@ def remove_files(user, folder):
     filepath = f'C:/Users/{user}/{folder}'
     dir_file_list = os.listdir(filepath)
     for i in dir_file_list:
-        print(i, end=" ")
+        print(i, end=", ")
 
     print("\nThe files above will be permanently excluded. Do you want to proceed?Y/n")
     input_user = str(input("->"))
     if input_user == 'Y' or 'y':
         for i in dir_file_list:
-            print(i)
-            os.remove(filepath+f"/{i}")
+            #print(filepath+f"/{i}")
+            if os.path.isfile(filepath+f"/{i}"):
+                os.remove(filepath+f"/{i}")
+                #print(os.path.isdir(filepath+f"/{i}"))
+            else:
+                while(file_inside_file):
+                    os.listdir(filepath+f"/{i}")
+                    
+
     elif input_user == 'N' or 'n':
         print("Canceling...")
         time.sleep(2.5)
@@ -41,8 +48,9 @@ def remove_files(user, folder):
         print("Invalid Number, please try again")
         remove_files(user,folder)
 
+
 def windows_remove_files_from_folder(user):
-    list_folders = ["Documents", "Images", "Downloads"]
+    list_folders = ["Documents", "Images", "Downloads/teste"]
     print("Which folder do you want to exclude the files?")
     print("[1] Documents")
     print("[2] Images")
@@ -57,8 +65,8 @@ def windows_remove_files_from_folder(user):
         remove_files(user,list_folders[int(input_user)-1])
     else:
         print("Invalid value, please try again")
-    
 
+    
 def main():
     welcome() 
     print("Select the option: ")
